made settings more secure
Passwords in the settings file are now encrypted using Windows Data
Protection (DPAPI). Only the machine the password was encrypted on will
be able to decrypt it. Also removed the ability to copy/paste from and
password field. Pressing 'Enter' in any settings field will cause a save
of the settings.